Move AllOf, AnyOf and OneOf
