+++ src/OpenAPI/Builder/APIBuilder.php
@@ -6,12 +6,12 @@ namespace Membrane\OpenAPI\Builder;
 
 use Membrane\Builder\Builder;
 use Membrane\OpenAPI;
-use Membrane\OpenAPI\Processor\AllOf;
-use Membrane\OpenAPI\Processor\AnyOf;
-use Membrane\OpenAPI\Processor\OneOf;
 use Membrane\OpenAPIReader\ValueObject\Valid\{Enum\Type, V30, V31};
 use Membrane\Processor;
+use Membrane\Processor\AllOf;
+use Membrane\Processor\AnyOf;
 use Membrane\Processor\Field;
+use Membrane\Processor\OneOf;
 use Membrane\Validator\Type\IsNull;
 use Membrane\Validator\Utility;
