+++ tests/OpenAPI/Builder/ObjectsTest.php
@@ -9,17 +9,17 @@ use Membrane\OpenAPI\Builder\Numeric;
 use Membrane\OpenAPI\Builder\Objects;
 use Membrane\OpenAPI\Builder\Strings;
 use Membrane\OpenAPI\Builder\TrueFalse;
-use Membrane\OpenAPI\Processor\AnyOf;
-use Membrane\OpenAPI\Processor\OneOf;
 use Membrane\OpenAPI\Specification;
 use Membrane\OpenAPIReader\ValueObject\Partial;
 use Membrane\OpenAPIReader\ValueObject\Valid\{Identifier, V30, V31};
 use Membrane\OpenAPIReader\ValueObject\Value;
 use Membrane\Processor;
+use Membrane\Processor\AnyOf;
 use Membrane\Processor\BeforeSet;
 use Membrane\Processor\DefaultProcessor;
 use Membrane\Processor\Field;
 use Membrane\Processor\FieldSet;
+use Membrane\Processor\OneOf;
 use Membrane\Validator\Collection\Contained;
 use Membrane\Validator\Collection\Count;
 use Membrane\Validator\FieldSet\FixedFields;
