+++ tests/OpenAPI/Builder/OpenAPIResponseBuilderTest.php
@@ -12,9 +12,6 @@ use Membrane\OpenAPI\Builder\OpenAPIResponseBuilder;
 use Membrane\OpenAPI\Exception\CannotProcessOpenAPI;
 use Membrane\OpenAPI\Exception\CannotProcessSpecification;
 use Membrane\OpenAPI\ExtractPathParameters\PathMatcher;
-use Membrane\OpenAPI\Processor\AllOf;
-use Membrane\OpenAPI\Processor\AnyOf;
-use Membrane\OpenAPI\Processor\OneOf;
 use Membrane\OpenAPI\Specification\APISchema;
 use Membrane\OpenAPI\Specification\Arrays;
 use Membrane\OpenAPI\Specification\Numeric;
@@ -26,10 +23,13 @@ use Membrane\OpenAPI\Specification\TrueFalse;
 use Membrane\OpenAPIReader\MembraneReader;
 use Membrane\OpenAPIReader\OpenAPIVersion;
 use Membrane\Processor;
+use Membrane\Processor\AllOf;
+use Membrane\Processor\AnyOf;
 use Membrane\Processor\BeforeSet;
 use Membrane\Processor\Collection;
 use Membrane\Processor\Field;
 use Membrane\Processor\FieldSet;
+use Membrane\Processor\OneOf;
 use Membrane\Result\FieldName;
 use Membrane\Result\Message;
 use Membrane\Result\MessageSet;
