+++ tests/OpenAPI/Builder/StringsTest.php
@@ -7,12 +7,12 @@ namespace Membrane\Tests\OpenAPI\Builder;
 use Membrane\Filter\String\ToUpperCase;
 use Membrane\OpenAPI\Builder\APIBuilder;
 use Membrane\OpenAPI\Builder\Strings;
-use Membrane\OpenAPI\Processor\AnyOf;
 use Membrane\OpenAPI\Specification;
 use Membrane\OpenAPIReader\ValueObject\Partial;
 use Membrane\OpenAPIReader\ValueObject\Valid\{Identifier, V30};
 use Membrane\OpenAPIReader\ValueObject\Value;
 use Membrane\Processor;
+use Membrane\Processor\AnyOf;
 use Membrane\Processor\Field;
 use Membrane\Validator\Collection\Contained;
 use Membrane\Validator\String\DateString;
