+++ tests/Processor/DefaultProcessorTest.php
@@ -7,12 +7,12 @@ namespace Membrane\Tests\Processor;
 use Generator;
 use Membrane\Filter;
 use Membrane\Filter\Type\ToFloat;
-use Membrane\OpenAPI\Processor\AllOf;
-use Membrane\OpenAPI\Processor\AnyOf;
-use Membrane\OpenAPI\Processor\OneOf;
 use Membrane\Processor;
+use Membrane\Processor\AllOf;
+use Membrane\Processor\AnyOf;
 use Membrane\Processor\DefaultProcessor;
 use Membrane\Processor\Field;
+use Membrane\Processor\OneOf;
 use Membrane\Result\FieldName;
 use Membrane\Result\Message;
 use Membrane\Result\MessageSet;
